Sort people list by group and write test for mapper function
